Ajustes a sweetAlert, permisos de reportes y filtrado de respuestas de api

+++ database/migrations/2025_01_10_160750_create_kardexes_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('kardexes', function (Blueprint $table) {
             $table->id();
             $table->foreignId('producto')->references('id')->on('productos');
-            $table->decimal('cantidad');
+            $table->integer('cantidad');
             $table->integer('accion'); //1 entrada 2 salida
             //$table->decimal('stockInicial');
             $table->foreignId('inventario')->references('id')->on('inventarios');
